debt verify added

+++ src/main/java/ge/bog/terminal/domain/Payment.java
@@ -1,7 +1,9 @@
 package ge.bog.terminal.domain;
 
+import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -9,7 +11,7 @@ public class Payment {
     private Long terminalId;
     private Long providerId;
     private String abonentCode;
-    private Long paymentAmount;
+    private BigDecimal paymentAmount;
     private String status;
     private LocalDateTime createTime;
 }
